Success/Failure Response for SMC and PGH reports
Uses error field in the response to contain a message for the client to display to the user in the event a potential action fails.
I.e. Too many reports submitted within a short time span or cause of swear word filtering.

Also changed the setup of models and schemas to resolve issues with setting up the server and database.

Changed rate limit for smell report submission to be one report every 5 minutes.